remove onDelete prop from ModComponentActionMenu
