+++ src/pageEditor/modListingPanel/DraftModComponentListItem.tsx
@@ -45,10 +45,6 @@ import {
 } from "@/pageEditor/store/editor/editorSelectors";
 import ModComponentActionMenu from "@/pageEditor/modListingPanel/ModComponentActionMenu";
 import useClearModComponentChanges from "@/pageEditor/hooks/useClearModComponentChanges";
-import {
-  useRemoveModComponentFromStorage,
-  DELETE_STARTER_BRICK_MODAL_PROPS,
-} from "@/pageEditor/hooks/useRemoveModComponentFromStorage";
 import { inspectedTab } from "@/pageEditor/context/connection";
 import { StarterBrickTypes } from "@/types/starterBrickTypes";
 
@@ -85,7 +81,6 @@ const DraftModComponentListItem: React.FunctionComponent<
   const isDirty = useSelector(
     selectModComponentIsDirty(modComponentFormState.uuid),
   );
-  const removeModComponentFromStorage = useRemoveModComponentFromStorage();
   const isButton =
     modComponentFormState.starterBrick.definition.type ===
     StarterBrickTypes.BUTTON;
@@ -99,12 +94,6 @@ const DraftModComponentListItem: React.FunctionComponent<
   }, []);
 
   const clearModComponentChanges = useClearModComponentChanges();
-
-  const deleteModComponent = async () =>
-    removeModComponentFromStorage({
-      modComponentId: modComponentFormState.uuid,
-      showConfirmationModal: DELETE_STARTER_BRICK_MODAL_PROPS,
-    });
 
   const onClearChanges = async () =>
     clearModComponentChanges({ modComponentId: modComponentFormState.uuid });
@@ -167,10 +156,10 @@ const DraftModComponentListItem: React.FunctionComponent<
         </span>
       )}
       <ModComponentActionMenu
+        modComponentFormState={modComponentFormState}
         isActive={isActive}
         isDirty={isDirty}
         labelRoot={getLabel(modComponentFormState)}
-        onDelete={deleteModComponent}
         onDuplicate={async () => {
           dispatch(
             // Duplicate the mod component in the same mod
